Fixed an issue with compiling files that haven't been permanently saved due to the files having a .bak extension throwing additional errors from the compiler. Now using .bak.nxc instead of .nxc.bak for this reason.

+++ jbrick/src/com/jbricx/actions/SaveAsAction.java
@@ -33,12 +33,12 @@ public class SaveAsAction extends AbstractAction {
 
 		// Check and see if it was previously saved as a backup
 		if (currDoc.getFileName() != null
-				&& currDoc.getFileName().endsWith(".bak")) {
+				&& currDoc.getFileName().endsWith(".bak.nxc")) {
 			String fname = currDoc.getFileName();
 			ActionControlClass.saveFile(getManager().getTabFolder()
 					.getSelection(), true, getManager(), getManager()
 					.getWorkspacePath());
-			if (!currDoc.getFileName().endsWith(".bak")) {
+			if (!currDoc.getFileName().endsWith(".bak.nxc")) {
 				// File was successfully saved, cleanup the temporary file
 				File f = new File(fname);
 				f.delete();
